+++ .dumirc.ts
@@ -1,7 +1,6 @@
 import { defineConfig } from 'dumi';
 
-// const publicPath = process.env.NODE_ENV === 'development' ? '/' : '/ld/o-orange/';
-const publicPath = process.env.NODE_ENV === 'development' ? '/' : '/o-orange/';
+const publicPath = process.env.NODE_ENV === 'development' ? '/' : process.env.PUBLIC_PATH;
 
 export default defineConfig({
   publicPath,