+++ src/constant/default.js
@@ -1,5 +1,0 @@
-"use strict";
-exports.__esModule = true;
-exports.DEFAULT_PLACEHOLDER = exports.DEFAULT_PRECISION = void 0;
-exports.DEFAULT_PRECISION = 2;
-exports.DEFAULT_PLACEHOLDER = '--';