+++ src/constant/default.ts
@@ -1,3 +1,0 @@
-export const DEFAULT_PRECISION = 2;
-
-export const DEFAULT_PLACEHOLDER = '--';