+++ src/orange.ts
@@ -1,4 +1,10 @@
 // see https://mikemcl.github.io/big.js/#rm
+export enum Lang {
+  ZH_CN = 'zh-CN',
+  ZH_TW = 'zh-TW',
+  EN_US = 'en-US'
+}
+
 export enum RoundingMode {
   roundDown,
   roundHalfUp,
@@ -7,5 +13,9 @@ export enum RoundingMode {
 }
 
 export default new class Orange {
-  RM = RoundingMode.roundHalfEven
+  RM = RoundingMode.roundHalfEven;
+  lang = Lang.EN_US;
+  precision = 2;
+  placeholder = '--';
+  ignoreIntegerPrecision = true;
 }();