+++ src/toPercent.ts
@@ -1,5 +1,4 @@
-import { DEFAULT_PRECISION, DEFAULT_PLACEHOLDER } from './constant/default';
-
+import orange from "./orange";
 import toFixed from './toFixed';
 
 interface IOption {
@@ -12,19 +11,14 @@ interface IOption {
  * Convert numbers to percentage
  *
  * @since 0.0.1
- * @param {number} num The number to convert
- * @param {Object} [option = {}]
- * @param {string} [option.placeholder = '--'] Replace string when targetNum is NaN or not number
- * @param {number} [option.precision = 2] The length to Keep
- * @param {number} [option.multiply = 1] 
- * @returns {string}
+ *
  *
  */
 function toPercent(
   num: number,
   option: IOption = {},
 ): string {
-  const { placeholder = DEFAULT_PLACEHOLDER, multiply = 1, precision = DEFAULT_PRECISION } = option;
+  const { placeholder = orange.placeholder, multiply = 1, precision = orange.precision } = option;
   const pureNum: number = Number(num);
 
   if (isNaN(pureNum)) return placeholder;