+++ src/toSlice.ts
@@ -1,8 +1,7 @@
-import { DEFAULT_PRECISION, DEFAULT_PLACEHOLDER } from './constant/default';
 import Big from 'big.js';
 import toFixed from './toFixed';
 import isTrue from './isTrue';
-import { RoundingMode } from './orange';
+import orange, { RoundingMode } from './orange';
 
 interface IOption {
   placeholder?: string;
@@ -14,20 +13,13 @@ interface IOption {
  * Slice decimal
  *
  * @since 3.0.3
- * @param {number} num The number to convert
- * @param {Object} [option = {}]
- * @param {string} [option.placeholder = '--'] Replace string when targetNum is NaN or not number
- * @param {number} [option.precision = 2] The length to slice
- * @param {boolean} [option.ignoreIntegerPrecision = false] if target is inter, ignore toFixed
- * @returns {string}
- * @Examples
  *
  */
 function toSlice(
   num: number | string,
   option: IOption = {},
 ): string {
-  const { placeholder = DEFAULT_PLACEHOLDER, precision = DEFAULT_PRECISION, ignoreIntegerPrecision = false } = option;
+  const { placeholder = orange.placeholder, precision = orange.precision, ignoreIntegerPrecision = false } = option;
   if (!isTrue(num)) return placeholder;
   const pureNum = Number(num);