+++ docs/CHANGELOG/changelog.md
@@ -1,8 +1,25 @@
 ---
 nav:
   title: Changelogs
-  order: 3
+  order: 4
 ---
+
+## v4
+
+### v4.1.0
+
+`2024-04-11`
+
+- docs: improve the quality of docs。
+- docs: add FAQs.
+- feat: support global config of `Lang`/`precision`/`placeholder`/`ignoreIntegerPrecision`.
+- test: add more test unit.
+
+### v4.0.0
+
+`2024-04-10`
+
+- chore: remove `parseUrl` to support SSR。
 
 ## v3